amélioration d'entite et de personnages

+++ lib/entite/attribut_entite.h
@@ -4,5 +4,5 @@ int w,h,w_hitbox,h_hitbox, nbTextures;
 SDL_Texture ** textures;
 chunk_t * chunk;
 salle_t * salle;
-int vitesse_x,vitesse_y;
+float vitesse_x,vitesse_y;
 booleen_t en_vie;
