+++ lib/entite/attribut_perso.h
@@ -1,5 +1,5 @@
 #include "attribut_entite.h"
-int vit_attack;
+float vit_attack;
 int degats;
 int vie;
-int vitesse_max;
+float vitesse_max;
